backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
django-celery-beat==2.6.0
drf-spectacular==0.27.2

psycopg2-binary==2.9.9
redis==5.0.3
celery==5.3.6

Pillow==10.3.0
python-decouple==3.8
gunicorn==22.0.0

requests==2.31.0
python-telegram-bot==21.2

openpyxl==3.1.2
python-dateutil==2.9.0
phonenumbers==8.13.36